% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_server.r
\name{analysisServer}
\alias{analysisServer}
\title{Analysis Server Module}
\usage{
analysisServer(id, data_obj)
}
\arguments{
\item{id}{Module namespace.}

\item{data_obj}{A reactive returning the processed data.}
}
\value{
Reactive simulation results.
}
\description{
Server logic for simulation and analysis tab.
}
