% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisUI.r
\name{analysisUI}
\alias{analysisUI}
\title{Analysis UI Module}
\usage{
analysisUI(id)
}
\arguments{
\item{id}{Module namespace.}
}
\value{
Shiny UI element.
}
\description{
UI for simulation and analysis tab. Allows effect size and spatial perturbation simulations.
}
