% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyDesignObject.R
\name{createDesignObject}
\alias{createDesignObject}
\title{Create shinyDesign object}
\usage{
createDesignObject(count_matrix, loc)
}
\arguments{
\item{count_matrix}{A gene expression count \code{matrix}}

\item{loc}{A \code{data.frame} containing the spatial coordinates and domain information for spots, corresponding to columns named as 'x', 'y', and 'domain'}
}
\value{
Returns a \code{shinyDesign} object
}
\description{
Create shinyDesign object
}
\examples{

## Create a shinyDesign object
toyDATA  <- createDesignObject(count_matrix = toyData$toyCount, loc = toyData$loc)

}
