% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInputServer.r
\name{dataInputServer}
\alias{dataInputServer}
\title{Data Input Server Module}
\usage{
dataInputServer(id, reference_data_paths)
}
\arguments{
\item{id}{A character string specifying the Shiny module namespace.}

\item{reference_data}{A list of reference datasets (optional).}
}
\value{
A reactive object containing the final processed data.
}
\description{
Server logic for the data input tab. Handles file reading, clustering, and processing.
}
