% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput_ui.r
\name{dataInputUI}
\alias{dataInputUI}
\title{Data Input UI Module}
\usage{
dataInputUI(id)
}
\arguments{
\item{id}{A character string specifying the Shiny module namespace.}
}
\value{
A Shiny UI element (sidebarLayout) for data input.
}
\description{
UI for the data input tab. Allows users to choose reference data or upload their own SpaceRanger output.
}
