% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams_GP_parallel.r
\name{estimateParams_GP_parallel}
\alias{estimateParams_GP_parallel}
\title{Estimate the parameters for Poisson Gaussian process model}
\usage{
estimateParams_GP_parallel(spadesign, nIter, nBurnin, covType, save = FALSE)
}
\arguments{
\item{spadesign}{A \code{spaDesign} object}

\item{nIter}{Number of MCMC iterations}

\item{nBurnin}{Number of MCMC iterations to exclude as burn-in period}

\item{covType}{Integer, specifying the covariance kernel, either exponential (default, \code{covType} = 1) or Gaussian kernel (\code{covType} = 0, a.k.a squared exponential, radial basis function kernel)}

\item{save}{Logical value, to save the posterior samples for the parameters (\code{save} = TRUE) or not (default, \code{save} = FALSE)}
}
\value{
Returns a \code{spaDesign} object with updated parameters
}
\description{
Estimate the parameters for Poisson Gaussian process model
}
\examples{
# Example usage:
# toyDATA <- createspadesignObject(count_matrix = toyData$toyCount, loc = toyData$loc)
# result <- estimateParams_GP(spadesign = toyDATA, nIter = 1000, nBurnin = 200, covType = 1, save = TRUE)
}
