% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParamsFG.r
\name{estimateParamsFG}
\alias{estimateParamsFG}
\title{Estimate the parameters for Fisher-Gaussian kernel mixture model}
\usage{
estimateParamsFG(spadesign, n_iter = 5000)
}
\arguments{
\item{spadesign}{A \code{spaDesign} object}

\item{n_iter}{number of MCMC iterations}
}
\description{
Estimate the parameters for Fisher-Gaussian kernel mixture model
}
\examples{
# Assuming spadesign is a valid spaDesign object
# result <- estimateParamsFG(spadesign, n_iter = 2000)
}
