% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParamsPoissonGP.r
\name{estimateParamsPoissonGP}
\alias{estimateParamsPoissonGP}
\title{Estimate Parameters for Poisson Gaussian Process Model}
\usage{
estimateParamsPoissonGP(
  spadesign,
  n_iter = 50000,
  n_burn = 30000,
  cov_type = 1,
  save = FALSE
)
}
\arguments{
\item{spadesign}{A \code{spaDesign} object}

\item{n_iter}{Number of MCMC iterations}

\item{n_burn}{Number of MCMC iterations to exclude as burn-in period}

\item{cov_type}{Integer specifying the covariance kernel (0 for Gaussian, 1 for Exponential)}

\item{save}{Logical value to save the posterior samples for the parameters (default = FALSE)}
}
\value{
Updated \code{spaDesign} object with parameter estimates and optionally posterior samples
}
\description{
Estimate Parameters for Poisson Gaussian Process Model
}
\examples{
## Example usage
toyDATA <- createSpaDesignObject(count_matrix = toyData$toyCount, loc = toyData$loc)
toyDATA <- estimateParamsPoissonGP(toyDATA, n_iter = 1000, n_burn = 200, cov_type = 1, save = FALSE)

}
