% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_NNGP_parallel.r
\name{estimation_NNGP_parallel}
\alias{estimation_NNGP_parallel}
\title{Estimate Parameters for Gaussian Process Model (Parallel Version)}
\usage{
estimation_NNGP_parallel(
  spadesign,
  n_neighbors = 10,
  order = "AMMD",
  verbose = FALSE,
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{spadesign}{A \code{spaDesign} object containing spatial coordinates and expression values}

\item{n_neighbors}{Number of nearest neighbors for NNGP fitting (default = 10)}

\item{order}{Ordering scheme for coordinates ('AMMD' or 'Sum_coords', default = 'AMMD')}

\item{verbose}{Whether to display progress messages (default = FALSE)}

\item{n_cores}{Number of cores to use for parallel processing (default = parallel::detectCores() - 1)}
}
\value{
Updated \code{spaDesign} object with parameter estimates
}
\description{
Fits a Nearest-Neighbor Gaussian Process (NNGP) model to estimate spatial parameters
for each domain-informative gene using parallel processing.
}
