% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data}
\format{
A list of the following components:
\describe{
\item{count}{A gene expression count matrix with 73 rows and 3611 columns}
\item{loc}{A dataframe with 3611 rows and 3 columns, corresponding to the spatial coordinates as well as the domain information for spots}
}
}
\source{
The example data is creased based on the SpatialLIBD data (sample ID: 151673).
}
\usage{
exampleData
}
\description{
This is an example dataset used to demonstrate the functionality of the package
}
\keyword{datasets}
