% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_main.R
\name{FG_EM}
\alias{FG_EM}
\title{EM algorithm for simplified Fisher-Gaussian model}
\usage{
FG_EM(x, M, iter_max = 500, tol = 0.1)
}
\arguments{
\item{x}{Numeric matrix of coordinates (n*d)}

\item{M}{Number of mixture components(clusters) (default: 5)}

\item{iter_max}{Maximum EM iterations (default: 1000)}

\item{tol}{Convergence tolerance (default: 1e-1)}
}
\description{
EM algorithm for simplified Fisher-Gaussian model
}
