% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_main.r
\name{FG_EM_with_criteria}
\alias{FG_EM_with_criteria}
\title{Fit FG-EM model and compute model selection criteria}
\usage{
FG_EM_with_criteria(x, M, iter_max, tol)
}
\arguments{
\item{x}{Numeric matrix of coordinates (n x d)}

\item{M}{Integer. Number of mixture components}

\item{iter_max}{Maximum EM iterations (default = 500)}

\item{tol}{Convergence tolerance for EM algorithm (default = 1e-1)}
}
\value{
A list containing:
\item{model}{List returned by FG_EM, including fitted parameters and log-likelihoods}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{n_parameters}{Number of estimated parameters in the model}
\item{converged_iter}{Number of EM iterations until convergence}
}
\description{
Runs the Fisher-Gaussian EM algorithm for a given number of clusters and computes
AIC and BIC for model selection.
}
\keyword{internal}
