% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParamsFG.r
\name{fgFit_single}
\alias{fgFit_single}
\title{Fit Fisher-Gaussian kernel mixture model for the spatial coordinates of a given tissue domain}
\usage{
fgFit_single(loc, n_iter)
}
\arguments{
\item{loc}{data frame. The normalized 2-d spatial coordinates of a spatial domain}

\item{n_iter}{number of MCMC iterations}
}
\value{
The fitted Fisher-Gaussian kernel mixture model
}
\description{
The Fisher-Gaussian kernel mixture model was proposed by M Mukhopadhyay et al., J R Stat Soc Series B Stat Methodol. 2020.
The model was originally developed for density estimation. Here, it is used to model the spatial distribution of spots from a tissue domain.
}
\details{
Fit Fisher-Gaussian kernel mixture for a single tissue domain
}
\examples{
# Assuming loc is a data frame with normalized coordinates
result <- fgFit_single(loc, n_iter = 2000)
}
