% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.r
\name{geneSummary}
\alias{geneSummary}
\title{calculate the fold change for each gene across spatial domains}
\usage{
geneSummary(count_matrix, loc)
}
\arguments{
\item{count_matrix}{A gene expression count matrix.}

\item{loc}{A dataframe containing the spatial coordiantes as well as domain information for each spot}
}
\value{
A list with fold change for each domain per gene
}
\description{
calculate the fold change for each gene across spatial domains
}
