% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_main.R
\name{initial_kmeans}
\alias{initial_kmeans}
\title{Initialize parameters for FG-EM algorithm using k-means}
\usage{
initial_kmeans(x, M)
}
\arguments{
\item{x}{Numeric matrix of coordinates}

\item{M}{Number of clusters}
}
\value{
A list containing initialized parameters:
\item{pi_hat}{Initial mixture weights}
\item{c_hat}{Initial cluster centers}
\item{r_hat}{Initial radii}
\item{sigma_sq}{Initial residual variance}
\item{phi_hat}{Initial mean directions}
\item{tau_hat}{Initial angular spreads}
}
\description{
Computes initial estimates of mixture weights, cluster centers, radii, angular directions,
angular spread, and residual variance for the Fisher-Gaussian EM algorithm.
}
\keyword{internal}
