% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGEM_main.r
\name{M_step}
\alias{M_step}
\title{M-step of the Fisher-Gaussian EM algorithm}
\usage{
M_step(x, W, y_expect_list, c_hat, r_hat, phi_hat, tau_hat)
}
\arguments{
\item{x}{Numeric matrix of coordinates (n x d)}

\item{W}{Posterior cluster probabilities (n x M)}

\item{y_expect_list}{List of expected normalized vectors for each cluster}

\item{c_hat}{Current cluster centers (M x d)}

\item{r_hat}{Current cluster radii}

\item{phi_hat}{Current mean directions (M x d)}

\item{tau_hat}{Current angular spreads}
}
\value{
A list with updated parameters:
\item{pi_hat}{Updated mixture weights}
\item{c_hat}{Updated cluster centers}
\item{r_hat}{Updated radii}
\item{phi_hat}{Updated mean directions}
\item{tau_hat}{Updated angular spreads}
\item{sigma_sq}{Updated Gaussian noise}
}
\description{
Updates mixture weights, cluster centers, radii, mean directions, angular spreads, and Gaussian noise
based on the current E-step results.
}
\keyword{internal}
