% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{mean_out}
\alias{mean_out}
\title{Calculate the Mean for Outside Domain Spots}
\usage{
mean_out(counts, gene, spot_idx)
}
\arguments{
\item{counts}{A matrix where rows represent genes and columns represent spots.}

\item{gene}{A character string representing the gene of interest.}

\item{spot_idx}{Numeric vector indicating the indices of spots within the domain.}
}
\value{
A numeric value representing the mean of the lower half of counts outside the domain for the given gene.
}
\description{
This function calculates the mean of the lower half of counts outside the domain for a given gene.
}
