% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{mean_out_refactored}
\alias{mean_out_refactored}
\title{(Refactored) Calculate Mean for Outside Domain Spots}
\usage{
mean_out_refactored(gene_count_row, spot_idx)
}
\arguments{
\item{gene_count_row}{A numeric vector of counts for a single gene.}

\item{spot_idx}{Numeric vector indicating the indices of spots within the domain.}
}
\value{
A numeric value representing the mean of the lower half of counts outside the domain.
}
\description{
(Refactored) Calculate Mean for Outside Domain Spots
}
