% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_NNGP.r
\name{nnGPfit}
\alias{nnGPfit}
\title{Fit a Nearest-Neighbor Gaussian Process(NNGP) model
The internal function that fits an NNGP model using the BRISC package}
\usage{
nnGPfit(
  spatial_coords,
  logCount,
  X = NULL,
  n_neighbors = 10,
  order = "AMMD",
  verbose = FALSE
)
}
\arguments{
\item{spatial_coords}{Matrix of spatial coordinates}

\item{logCount}{vector of log-transformed gene expression counts}

\item{X}{Optional design matrix of covariates}

\item{n_neighbors}{Number of nearest neighbors for fitting NNGP model with BRISC. Default = 10.}

\item{verbose}{Whether to display process messages (default = FALSE)}

\item{order:}{Ordering scheme to use for ordering coordinates with BRISC.('AMMD' or 'Sum_coords', default = 'AMMD'). See BRISC documentation for details.}
}
\value{
BRISC fitted model objet
}
\description{
Fit a Nearest-Neighbor Gaussian Process(NNGP) model
The internal function that fits an NNGP model using the BRISC package
}
\keyword{internal}
