% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{paramsFG,shinyDesign-method}
\alias{paramsFG,shinyDesign-method}
\title{Access fitted Fisher-Gaussian kernel model parameters from a shinyDesign object}
\usage{
\S4method{paramsFG}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
A list of estimated parameters from fitted Fisher-Gaussian kernel mixture model
}
\description{
Access fitted Fisher-Gaussian kernel model parameters from a shinyDesign object
}
