% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{paramsFG,spaDesign-method}
\alias{paramsFG,spaDesign-method}
\title{Access fitted Fisher-Gaussian kernel model parameters from a spaDesign object}
\usage{
\S4method{paramsFG}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
A list of estimated parameters from fitted Fisher-Gaussian kernel mixture model
}
\description{
Access fitted Fisher-Gaussian kernel model parameters from a spaDesign object
}
