% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paramsFG}
\alias{paramsFG}
\title{Access estimated parameter from the fitted Fisher-Gaussian kernel model}
\usage{
paramsFG(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns a list of estimated parameters from fitted Fisher-Gaussian kernel mixture model
}
\description{
Access estimated parameter from the fitted Fisher-Gaussian kernel model
}
\examples{
paramsFG(shinyDesign)
}
