% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{paramsGP,shinyDesign-method}
\alias{paramsGP,shinyDesign-method}
\title{Access fittedNNGP model parameters from a shinyDesign object}
\usage{
\S4method{paramsGP}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
A list of estimated parameters from fitted NNGP model
}
\description{
Access fittedNNGP model parameters from a shinyDesign object
}
