% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paramsGP}
\alias{paramsGP}
\title{Access estimated parameter from the fitted Gaussian process model}
\usage{
paramsGP(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns a list of estimated parameters from fitted Gaussian process model
}
\description{
Access estimated parameter from the fitted Gaussian process model
}
\examples{
paramsGP(shinyDesign)
}
