% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{paramsPoissonGP,spaDesign-method}
\alias{paramsPoissonGP,spaDesign-method}
\title{Access fitted Poisson-Gaussian model parameters from a spaDesign object}
\usage{
\S4method{paramsPoissonGP}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
A list of estimated parameters from fitted Poisson Gaussian process model
}
\description{
Access fitted Poisson-Gaussian model parameters from a spaDesign object
}
