% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paramsPoissonGP}
\alias{paramsPoissonGP}
\title{Access estimated parameter from the fitted Poisson Gaussian process model}
\usage{
paramsPoissonGP(x)
}
\arguments{
\item{x}{\code{spaDesign} object}
}
\value{
Returns a list of estimated parameters from fitted Poisson Gaussian process model
}
\description{
Access estimated parameter from the fitted Poisson Gaussian process model
}
\examples{
paramsPoissonGP(spaDesign)
}
