% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParamsPoissonGP.r,
%   R/estimateParams_GP_parallel.r
\name{PoissonGP_multi}
\alias{PoissonGP_multi}
\title{Fit Poisson Gaussian Process Model for Multiple Domain-Informatives Genes}
\usage{
PoissonGP_multi(coords, counts, n_iter, n_burn, cov_type)

PoissonGP_multi(coords, counts, n_iter, n_burn, cov_type)
}
\arguments{
\item{coords}{Matrix of 2-dimensional locations}

\item{counts}{Matrix of gene expression counts (rows = genes, columns = locations)}

\item{n_iter}{Number of MCMC iterations}

\item{n_burn}{Number of MCMC iterations to exclude as burn-in period}

\item{cov_type}{Integer specifying the covariance kernel (0 for Gaussian, 1 for Exponential)}
}
\value{
A list of results for each gene including MCMC samples and computation time

A list of results for each gene including MCMC samples and computation time
}
\description{
Fit Poisson Gaussian Process Model for Multiple Domain-Informatives Genes

Fit Poisson Gaussian Process Model for Multiple Domain-Informatives Genes
}
