% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysisSpatial.r
\name{powerAnalysisSpatial}
\alias{powerAnalysisSpatial}
\title{Power analysis for spatial data}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object with estimated model parameters}

\item{prop_range}{A numeric vector indicating the range of proportions (proportion of genes with undisturbed pattern)}

\item{seq_depth_range}{A numeric vector indicating the range of sequencing depth scaling factors}

\item{n_rep}{An integer indicating the number of replications for each scenario}

\item{conda_env_path}{Specify the path to the conda environment for spaGCN}

\item{python_script}{Specify the path to the python script}
}
\value{
A data frame containing the sequencing depth, proportion, total counts, and NMI
}
\description{
This function performs power analysis by simulating spatial data with different proportions of undisturbed patterns and sequencing depths.
}
\examples{
# spadesign <- ...
# prop_range <- c(0.1, 0.5, 1)
# seq_depth_range <- c(0.5, 1, 3, 5)
# n_rep <- 10
# results <- powerAnalysisSpatial(shinyDesign, prop_range, seq_depth_range, n_rep, conda_env_path) 
}
