% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refcolData,shinyDesign-method}
\alias{refcolData,shinyDesign-method}
\title{Access the reference column data from a shinyDesign object}
\usage{
\S4method{refcolData}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
The reference column data
}
\description{
Access the reference column data from a shinyDesign object
}
