% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refcolData,spaDesign-method}
\alias{refcolData,spaDesign-method}
\title{Access the reference column data from a spaDesign object}
\usage{
\S4method{refcolData}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
The reference column data
}
\description{
Access the reference column data from a spaDesign object
}
