% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{refcolData}
\alias{refcolData}
\title{Extract the reference colData}
\usage{
refcolData(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns the colData of reference data
}
\description{
Extract the reference colData
}
\examples{
refcolData(shinyDesign)
}
