% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refCounts,shinyDesign-method}
\alias{refCounts,shinyDesign-method}
\title{Access the reference count matrix from a shinyDesign object}
\usage{
\S4method{refCounts}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
The reference count matrix
}
\description{
Access the reference count matrix from a shinyDesign object
}
