% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refCounts,spaDesign-method}
\alias{refCounts,spaDesign-method}
\title{Access the reference count matrix from a spaDesign object}
\usage{
\S4method{refCounts}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
The reference count matrix
}
\description{
Access the reference count matrix from a spaDesign object
}
