% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{refCounts}
\alias{refCounts}
\title{Extract the reference count matrix}
\usage{
refCounts(x)
}
\arguments{
\item{x}{shinyDesign object}
}
\value{
Returns a reference count matrix
}
\description{
Extract the reference count matrix
}
\examples{
refCounts(shinyDesign)
}
