% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refrowData,shinyDesign-method}
\alias{refrowData,shinyDesign-method}
\title{Access the reference row data from a shinyDesign object}
\usage{
\S4method{refrowData}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
The reference row data
}
\description{
Access the reference row data from a shinyDesign object
}
