% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{refrowData,spaDesign-method}
\alias{refrowData,spaDesign-method}
\title{Access the reference row data from a spaDesign object}
\usage{
\S4method{refrowData}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
The reference row data
}
\description{
Access the reference row data from a spaDesign object
}
