% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{refrowData}
\alias{refrowData}
\title{Access reference rowData}
\usage{
refrowData(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns the rowData of reference data
}
\description{
Access reference rowData
}
\examples{
refrowData(shinyDesign)
}
