% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\name{shinyDesign-accessors}
\alias{shinyDesign-accessors}
\alias{refCounts}
\alias{refcolData}
\alias{refrowData}
\alias{simCounts}
\alias{simcolData}
\alias{simrowData}
\alias{paramsFG}
\alias{paramsGP}
\alias{topGenes}
\alias{refCounts,shinyDesign-method}
\alias{refcolData,shinyDesign-method}
\alias{refrowData,shinyDesign-method}
\alias{simCounts,shinyDesign-method}
\alias{simcolData,shinyDesign-method}
\alias{simrowData,shinyDesign-method}
\alias{paramsFG,shinyDesign-method}
\alias{paramsGP,shinyDesign-method}
\alias{topGenes,shinyDesign-method}
\title{Access slots of a shinyDesign object}
\usage{
refCounts(x)

refcolData(x)

refrowData(x)

simCounts(x)

simcolData(x)

simrowData(x)

paramsFG(x)

paramsGP(x)

topGenes(x)

\S4method{refCounts}{shinyDesign}(x)

\S4method{refcolData}{shinyDesign}(x)

\S4method{refrowData}{shinyDesign}(x)

\S4method{simCounts}{shinyDesign}(x)

\S4method{simcolData}{shinyDesign}(x)

\S4method{simrowData}{shinyDesign}(x)

\S4method{paramsFG}{shinyDesign}(x)

\S4method{paramsGP}{shinyDesign}(x)

\S4method{topGenes}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object.}
}
\value{
Slot content (matrix, data.frame, or list) depending on the accessor

Slot contents:
\itemize{
\item \code{refCounts} – reference count matrix
\item \code{refcolData} – reference column metadata
\item \code{refrowData} – reference row metadata
\item \code{simCounts} – simulated count matrix
\item \code{simcolData} – simulated column metadata
\item \code{simrowData} – simulated row metadata
\item \code{paramsFG} – Fisher–Gaussian mixture model parameters
\item \code{paramsGP} – Gaussian process/NNGP model parameters
\item \code{topGenes} – list of top selected genes
}
}
\description{
Access reference counts, simulated counts, metadata, model parameters, and top genes
stored in a \code{shinyDesign} object.

Access slots of a \code{shinyDesign} object, including reference
and simulated data, model parameters, and results.
}
\examples{
\dontrun{
sd <- new("shinyDesign",
          refCounts = matrix(1:6, nrow = 2),
          refcolData = data.frame(sample = 1:3),
          refrowData = data.frame(gene = c("A","B")),
          simCounts = matrix(7:12, nrow = 2),
          simcolData = data.frame(cell = 1:3),
          simrowData = data.frame(gene = c("A","B")),
          paramsFG = list(param1 = 1),
          paramsGP = list(param2 = 2),
          topGenes = list(c("gene1","gene2")))

refCounts(sd)
refcolData(sd)
refrowData(sd)
simCounts(sd)
simcolData(sd)
simrowData(sd)
paramsFG(sd)
paramsGP(sd)
topGenes(sd)
}
\dontrun{
obj <- new("shinyDesign",
           refCounts = matrix(1:6, nrow = 2),
           refcolData = data.frame(sample = 1:3),
           refrowData = data.frame(gene = c("A", "B")),
           simCounts = matrix(7:12, nrow = 2),
           simcolData = data.frame(cell = 1:3),
           simrowData = data.frame(gene = c("A", "B")))

refCounts(obj)
simCounts(obj)
paramsFG(obj)
}

}
