% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_options_and_data.r
\docType{package}
\name{shinyDesign2-package}
\alias{shinyDesign2}
\alias{shinyDesign2-package}
\title{shinyDesign2: Tools for sequencing-based spatial transcriptomics experimental design}
\description{
Provides tools for simulating sequencing-based spatial transcriptomics data to facilitate experimental designs. Includes integration with Seurat, Shiny modules for interactive analysis, and utility functions for performance evaluation.
}
\author{
\strong{Maintainer}: Juan Xie \email{juanxie.84@gmail.com}

}
\keyword{internal}
