% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{show,shinyDesign-method}
\alias{show,shinyDesign-method}
\title{Show method for shinyDesign objects}
\usage{
\S4method{show}{shinyDesign}(object)
}
\arguments{
\item{object}{A \code{shinyDesign} object}
}
\description{
Show method for shinyDesign objects
}
