% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.r
\name{show,spaDesign-method}
\alias{show,spaDesign-method}
\title{Show method for spaDesign objects}
\usage{
\S4method{show}{spaDesign}(object)
}
\arguments{
\item{object}{A \code{spaDesign} object}
}
\description{
Show method for spaDesign objects
}
