% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{simcolData,shinyDesign-method}
\alias{simcolData,shinyDesign-method}
\title{Access the simulated column data from a shinyDesign object}
\usage{
\S4method{simcolData}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
The simulated column data
}
\description{
Access the simulated column data from a shinyDesign object
}
