% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{simcolData,spaDesign-method}
\alias{simcolData,spaDesign-method}
\title{Access the simulated column data from a spaDesign object}
\usage{
\S4method{simcolData}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
The simulated column data
}
\description{
Access the simulated column data from a spaDesign object
}
