% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simcolData}
\alias{simcolData}
\title{Access colData for simulated data}
\usage{
simcolData(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns the colData of simulated data
}
\description{
Access colData for simulated data
}
\examples{
simcolData(shinyDesign)
}
