% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{simCounts,shinyDesign-method}
\alias{simCounts,shinyDesign-method}
\title{Access the simulated count matrix from a shinyDesign object}
\usage{
\S4method{simCounts}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
The simulated count matrix
}
\description{
Access the simulated count matrix from a shinyDesign object
}
