% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{simCounts,spaDesign-method}
\alias{simCounts,spaDesign-method}
\title{Access the simulated count matrix from a spaDesign object}
\usage{
\S4method{simCounts}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
The simulated count matrix
}
\description{
Access the simulated count matrix from a spaDesign object
}
