% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simCounts}
\alias{simCounts}
\title{Access simulated count matrix}
\usage{
simCounts(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns the simulated count matrix
}
\description{
Access simulated count matrix
}
\examples{
simCounts(shinyDesign)
}
