% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simrowData}
\alias{simrowData}
\title{Access rowData for simulated data}
\usage{
simrowData(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns the rowData of simulated data
}
\description{
Access rowData for simulated data
}
\examples{
simrowData(shinyDesign)
}
