% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{simulate_genecount_ES_refactored}
\alias{simulate_genecount_ES_refactored}
\title{(Refactored) Simulate expression count for a single gene}
\usage{
simulate_genecount_ES_refactored(
  SEED,
  coords_norm_sub,
  coords_norm_out,
  nnGP_fit,
  seqDepth_factor,
  ES_factor,
  gene_count_row,
  all_spot_names,
  spot_idx
)
}
\arguments{
\item{SEED}{Random seed for reproducibility}

\item{coords_norm_sub}{Normalized coordinates for spots within the domain}

\item{coords_norm_out}{Normalized coordinates for spots outside the domain}

\item{nnGP_fit}{ouutput from parameter estimation step}

\item{ES_factor}{Effect size factor}

\item{gene_count_row}{A numeric vector of counts for the gene of interest. # <-- CHANGED}

\item{all_spot_names}{A character vector of all spot names (from colnames). # <-- NEW}

\item{spot_idx}{Indices of spots within the domain}

\item{seq_depth_factor}{Sequencing depth factor}
}
\value{
Simulated counts for the gene
}
\description{
(Refactored) Simulate expression count for a single gene
}
