% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataEffectSize.r
\name{simulateDataEffectSize}
\alias{simulateDataEffectSize}
\title{simualte data with modified effect size and sequencing depth}
\usage{
simulateDataEffectSize(spadesign, seq_depth_factor, effect_size_factor, SEED)
}
\arguments{
\item{spadesign}{The \code{spadesign} object}

\item{seq_depth_factor}{A scaling factor for sequencing depth}

\item{effect_size_factor}{A scaling factor for effect size}

\item{SEED}{random seed for reproducibility}
}
\value{
A spatialdesign object with simulated count matrix
}
\description{
simualte data with modified effect size and sequencing depth
}
