% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataSpatial.r
\name{simulateDataSpatial}
\alias{simulateDataSpatial}
\title{simulate data with modified sequencing depth and dispersed pattern}
\usage{
simulateDataSpatial(spadesign, seq_depth_factor, SIGMA, SEED, prop)
}
\arguments{
\item{spadesign}{The \code{spadesign} object}

\item{seq_depth_factor}{A scaling factor for sequencing depth}

\item{SIGMA}{scaling factor for the \code{sigma} parameter from the FG model}

\item{SEED}{random seed for reproducibility}

\item{prop}{Proportion of genes with undisturbed expression pattern}
}
\value{
A \code{spadesign} object with simulated count matrix
}
\description{
simulate data with modified sequencing depth and dispersed pattern
}
