% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{simulation_EffectSize_refactored}
\alias{simulation_EffectSize_refactored}
\title{(Refactored) Simulate spatial transcriptomics data}
\usage{
simulation_EffectSize_refactored(
  shinyDesign,
  seq_depth_factor,
  effect_size_factor,
  SEED
)
}
\description{
(Refactored) Simulate spatial transcriptomics data
}
