% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{spaDesign-class}
\alias{spaDesign-class}
\title{Class \code{spaDesign}}
\value{
A \code{spaDesign} object.
}
\description{
The \code{spaDesign} class is used to store and manage data relevant to power analysis
within the \code{spaDesign} package. It holds both reference and simulated count data, associated meta data,
and optional parameters from fitted models.
}
\details{
The \code{spaDesign} class includes the following slots:
\describe{
\item{\code{refCounts}}{A slot for reference count data}
\item{\code{refcolData}}{A \code{data.frame} containing column-level metadata for reference counts.}
\item{\code{refrowData}}{A \code{data.frame} containing row-level metadata for reference counts.}
\item{\code{simCounts}}{A slot for simulated count data}
\item{\code{simcolData}}{A \code{data.frame} containing column-level metadata for simulated counts.}
\item{\code{simrowData}}{A \code{data.frame} containing row-level metadata for simulated counts.}
\item{\code{paramsGP}}{An optional list of parameters for Gaussian processes, or \code{NULL}.}
\item{\code{paramsFG}}{An optional list of parameters for Fisher Gaussian kernel mixture model, or \code{NULL}.}
\item{\code{selected_M_list_AIC}}{A slot for selected best number of cluster for FG mixture model based on AIC criteria.}
\item{\code{selected_M_list_BIC}}{A slot for selected best number of cluster for FG mixture model based on BIC criteria.}
\item{\code{topGenes}}{An optional list of top genes, or \code{NULL}.}
\item{\code{NMI}}{A slot for normalized mutual information}
}

This class provides a container for storing both observed and simulated
spatial transcriptomics data, along with results from downstream modeling.
}
\section{Slots}{

\describe{
\item{\code{refCounts}}{A matrix-like object (e.g., \code{Matrix::dgCMatrix}) of reference counts.}

\item{\code{refcolData}}{A \code{data.frame} containing column-level metadata for reference counts.}

\item{\code{refrowData}}{A \code{data.frame} containing row-level metadata for reference counts.}

\item{\code{simCounts}}{A matrix-like object of simulated counts.}

\item{\code{simcolData}}{A \code{data.frame} containing column-level metadata for simulated counts.}

\item{\code{simrowData}}{A \code{data.frame} containing row-level metadata for simulated counts.}

\item{\code{paramsGP}}{An optional list of parameters for Gaussian processes, or \code{NULL}.}

\item{\code{paramsFG}}{An optional list of parameters for Fisher Gaussian kernel mixture model, or \code{NULL}.}

\item{\code{selected_M_list_AIC}}{Best model selection results based on AIC.}

\item{\code{selected_M_list_BIC}}{Best model selection results based on BIC.}

\item{\code{topGenes}}{An optional list of top genes, or \code{NULL}.}

\item{\code{NMI}}{Normalized mutual information value (numeric or object).}
}}

