% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_setup.r
\docType{package}
\name{spaDesign-package}
\alias{spaDesign}
\alias{spaDesign-package}
\title{spaDesign: Tools for sequencing-based spatial transcriptomics experimental design}
\description{
Provides tools for simulating sequencing-based spatial transcriptomics data to facilitate experimental designs. Includes integration with Seurat, Shiny modules for interactive analysis, and utility functions for performance evaluation.
}
\author{
\strong{Maintainer}: Juan Xie \email{juanxie.84@gmail.com}

}
\keyword{internal}
