% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{topGenes,shinyDesign-method}
\alias{topGenes,shinyDesign-method}
\title{Access selected top genes for each domain from a shinyDesign object}
\usage{
\S4method{topGenes}{shinyDesign}(x)
}
\arguments{
\item{x}{A \code{shinyDesign} object}
}
\value{
A list of selected top genes
}
\description{
Access selected top genes for each domain from a shinyDesign object
}
