% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\name{topGenes,spaDesign-method}
\alias{topGenes,spaDesign-method}
\title{Access selected top genes for each domain from a spaDesign object}
\usage{
\S4method{topGenes}{spaDesign}(x)
}
\arguments{
\item{x}{A \code{spaDesign} object}
}
\value{
A list of selected top genes
}
\description{
Access selected top genes for each domain from a spaDesign object
}
