% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{topGenes}
\alias{topGenes}
\title{Access selected top genes for each domain}
\usage{
topGenes(x)
}
\arguments{
\item{x}{\code{shinyDesign} object}
}
\value{
Returns a list of selected top genes for each domain
}
\description{
Access selected top genes for each domain
}
\examples{
topGenes(shinyDesign)
}
