% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisServer _test.r, R/analysisServer.r,
%   R/analysisServer_test2.r
\name{analysisServer}
\alias{analysisServer}
\title{Analysis Server Module}
\usage{
analysisServer(id, data_obj)

analysisServer(id, data_obj)

analysisServer(id, data_obj)
}
\arguments{
\item{id}{Module namespace.}

\item{data_obj}{A reactive returning the processed data.}
}
\value{
Reactive simulation results.

Reactive simulation results.

Reactive simulation results.
}
\description{
Server logic for simulation and analysis tab.

Server logic for simulation and analysis tab.

Server logic for simulation and analysis tab.
}
