% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpaDesignObject.r
\name{createDesignObject}
\alias{createDesignObject}
\title{Create a spaDesign object}
\usage{
createDesignObject(count_matrix, loc)
}
\arguments{
\item{count_matrix}{A numeric matrix of gene expression count(genes x spots)}

\item{loc}{A \code{data.frame} containing the spatial coordinates and domain information for spots,
Required columns: 'x', 'y', and 'domain'.}
}
\value{
A \code{spaDesign} object
}
\description{
Creates a \code{spaDesign} object to store reference and simulated count data,
along with associated spatial metadata and optional model parameters.
}
\examples{
\dontrun{
sd <- createDesignObject(count_matrix = toyData$toyCount, loc = toyData$loc)
}
}
