% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_FGEM.r
\name{estimation_FGEM}
\alias{estimation_FGEM}
\title{Estimate the parameters for simplified Fisher-Gaussian kernel mixture model using EM algorithm}
\usage{
estimation_FGEM(shinyDesign, iter_max = 1000, M_candidates = 2:5, tol = 0.1)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object containing spatial data}

\item{iter_max}{Maximum number of EM iterations (default = 1000)}

\item{tol}{Convergence tolerance for EM algorithm (default = 1e-1)
return Updated \code{shinyDesign} object with Fisher-Gaussian parameter estimates}

\item{M}{Number of mixture components(clusters) (default = 5)}
}
\description{
Fits a Fisher-Gaussian kernel mixture model to spatial coordinates using the EM algorithm. This models the spatial distribution of spots within each tissue domain
}
\examples{
# Assuming shinyDesign is a valid shinyDesign object
# result <- estimation_FGEM(shinyDesign, iter_max = 1000, M =5, tol = 1e-1)
}
