% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_NNGP.r
\name{estimation_NNGP}
\alias{estimation_NNGP}
\title{Estimate the NNGP Parameters for domain-informative genes}
\usage{
estimation_NNGP(
  spaDesign,
  n_neighbors = 10,
  order = "AMMD",
  X = NULL,
  verbose = FALSE
)
}
\arguments{
\item{spaDesign}{A \code{spaDesign} object containing spatial coordinates in \code{refcolData}
and expression values in \code{refCounts}.}

\item{n_neighbors}{Number of nearest neighbors for NNGP fitting (default = 10)}

\item{order}{Ordering scheme for coordinates('AMMD' or 'Sum_coords', default = 'AMMD').
use 'Sum_coords' for non-visium data}

\item{X}{Optional design matrix of covariates}

\item{verbose}{Logical, whether to display progress messages (default = FALSE)}
}
\value{
Updated \code{spaDesign} object. The slot \code{paramsGP} is a nested list:
\code{paramsGP[[domain]][[gene]]} contains the BRISC NNGP fit for that gene.

Updated \code{spaDesign} object with \code{paramsGP} slot populated.
This is a nested list: \code{paramsGP[[domain]][[gene]]} contains the BRISC fit object.
}
\description{
Fits a Nearest-Neighbor Gaussian Process (NNGP) model for each domain-informative gene
using the BRISC package.
}
\details{
Coordinates are normalized to \link{0,1} range before fitting. For each domain, NNGP models
are fit to log-transformed expression (log(count + 1)) for all selected genes within that domain.
}
\examples{
\dontrun{
toyDATA <- createDesignObject(count_matrix = toyData$toyCount, 
                               loc = toyData$loc)
toyDATA <- featureSelection(toyDATA, logfc_cutoff = 0.7, 
                            mean_in_cutoff = 2, max_num_gene = 50, n_cores = 4)
toyDATA <- estimation_NNGP(toyDATA, n_neighbors = 10, order = 'AMMD')
}
}
