% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_NNGP.r
\name{estimation_NNGP}
\alias{estimation_NNGP}
\title{Estimate Parameters for Gaussian Process Model}
\usage{
estimation_NNGP(
  shinyDesign,
  n_neighbors = 10,
  ORDER = "AMMD",
  X = NULL,
  verbose = FALSE
)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object containing spatial coordiantes and expression values for selected domain-informative genes}

\item{n_neighbors}{Number of nearest neighbors for NNGP fitting (default = 10)}

\item{verbose}{Whether to display progress messages (default = FALSE)}

\item{order}{Ordering scheme for coordinates('AMMD' or 'Sum_coords', default = 'AMMD')}
}
\value{
Updated \code{shinyDesign} object with parameter estimates
}
\description{
Fits a Nearest-Neighbor Gaussian Process (NNGP) model to estimate spatial parameters
for each domain-informative gene.
}
\examples{
## Example usage
toyDATA <- createshinyDesignObject(count_matrix = toyData$toyCount, loc = toyData$loc)
toyDATA <- estimation_NNGP(toyDATA, n_neighbors = 10, order = 'AMMD', verbose = FALSE)

}
