% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{mean_in}
\alias{mean_in}
\title{Compute mean expression inside a domain}
\usage{
mean_in(SEED, coords_norm_sub, nnGP_fit)
}
\arguments{
\item{SEED}{Integer random seed for reproducibility}

\item{coords_norm_sub}{Data frame of normalized coordinates for spots within the domain.}

\item{nnGP_fit}{Fitted NNGP model object for the gene.}
}
\value{
Numeric vector of predicted expression levels for each spot within the domain
}
\description{
Uses a fitted NNGP model to predict expected expression levels for spots within a domain.
}
