% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{mean_in}
\alias{mean_in}
\title{This function simulates data based on fitted NNGP model}
\usage{
mean_in(SEED, coords_norm_sub, nnGP_fit)
}
\arguments{
\item{SEED}{Random seed for reproducibility}

\item{coords_norm_sub}{Normalized coordinates for spots within the domain}

\item{nnGP_fit}{ouutput from parameter estimation step}
}
\value{
A numeric vector representing the sampled expression levels within the domain
}
\description{
This function simulates data based on fitted NNGP model
}
