% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{mean_out}
\alias{mean_out}
\title{Compute mean expression for spots outside the domainO}
\usage{
mean_out(counts, gene, spot_idx)
}
\arguments{
\item{counts}{Matrix of gene expression counts (genes x spots).}

\item{gene}{Character string specifying the gene name.}

\item{spot_idx}{Numeric vector indicating the indices of spots within the domain.}
}
\value{
Numeric value representing the mean of the lower half of counts outside the domain for the given gene.
}
\description{
Calculates the mean of the lower half of expression counts for spots outside the domain for a given gene.
}
