% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysisEffectSize.r
\name{powerAnalysisEffectSize}
\alias{powerAnalysisEffectSize}
\title{Power analysis}
\usage{
powerAnalysisEffectSize(shinyDesign, es_range, seq_depth_range, n_rep)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object with estimated model parameters}

\item{es_range}{A numeric vector indicating the range of effect size scaling factors}

\item{seq_depth_range}{A numeric vector indicating the range of sequencing depth scaling factors}

\item{n_rep}{An integer indicating the number of replications for each scenario}

\item{conda_env_path}{Specify the path to the conda environment for spaGCN}

\item{python_script}{Specify the path to the python script}
}
\value{
A data frame containing the sequencing depth, effect size, total counts, and NMI
}
\description{
This function performs power analysis by simulating data with different effect sizes and sequencing depths.
}
\examples{
# spadesign <- ...
# es_range <- c(1, 2, 3)
# seq_depth_range <- c(0.5, 1, 3, 5)
# n_rep <- 10
# results <- powerAnalysisEffectSize(shinyDesign, es_range, seq_depth_range, n_rep, conda_env_path) 
}
