% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r,
%   R/simulation_EffectSize_test.r
\name{simulate_genecount_ES}
\alias{simulate_genecount_ES}
\title{Simulate expression count for a single gene}
\usage{
simulate_genecount_ES(
  SEED,
  coords_norm_sub,
  coords_norm_out,
  nnGP_fit,
  seqDepth_factor,
  ES_factor,
  counts,
  gene,
  spot_idx
)

simulate_genecount_ES(
  SEED,
  coords_norm_sub,
  coords_norm_out,
  nnGP_fit,
  seqDepth_factor,
  ES_factor,
  counts,
  gene,
  spot_idx
)
}
\arguments{
\item{SEED}{Random seed for reproducibility}

\item{coords_norm_sub}{Normalized coordinates for spots within the domain}

\item{coords_norm_out}{Normalized coordinates for spots outside the domain}

\item{nnGP_fit}{ouutput from parameter estimation step}

\item{counts}{Count matrix}

\item{gene}{Gene name}

\item{spot_idx}{Indices of spots within the domain}

\item{seq_depth_factor}{Sequencing depth factor}

\item{es_factor}{Effect size factor}
}
\value{
Simulated counts for the gene

Simulated counts for the gene
}
\description{
Simulate expression count for a single gene

Simulate expression count for a single gene
}
