% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_EffectSize.r
\name{simulation_EffectSize}
\alias{simulation_EffectSize}
\title{Simualte spatial transcriptomics data with modified effect size and sequencing depth}
\usage{
simulation_EffectSize(shinyDesign, seq_depth_factor, effect_size_factor, SEED)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object containing spatial coordiantes, gene expression and estimated parameters}

\item{seq_depth_factor}{Numeric scaling factor for sequencing depth (must be > 0 )}

\item{effect_size_factor}{Numeric scaling factor for effect size (must be > 0)}

\item{SEED}{random seed for reproducibility (integer)}
}
\value{
A \code{spatialdesign} object with simulated count matrix in the \code{simCounts} slot
}
\description{
Simualte spatial transcriptomics data with modified effect size and sequencing depth
}
\examples{
\dontrun{
simulated_data <- simulation_EffectSize(spadesign_obj, 
                                       seq_depth_factor = 1.5,
                                       effect_size_factor = 2.0,
                                       SEED = 123)
}
}
