% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_Spatial.r
\name{simulation_Spatial}
\alias{simulation_Spatial}
\title{Simulate data with modified sequencing depth and dispersion}
\usage{
simulation_Spatial(
  shinyDesign,
  selected_M_list = NULL,
  seq_depth_factor,
  SIGMA,
  SEED,
  prop
)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object containing spatial coordinates, gene expression,
and fitted GP/FG parameters.}

\item{selected_M_list}{Optional list of selected FG model indices for each domain. If \code{NULL},
the function will attempt to use \code{shinyDesign@selected_M_list_BIC}.}

\item{seq_depth_factor}{Numeric scaling factor for sequencing depth}

\item{SIGMA}{Numeric scaling factor for the \code{sigma} parameter from the FG model.}

\item{SEED}{Integer random seed for reproducibility.}

\item{prop}{Numeric, proportion of genes to keep undisturbed (between 0 and 1).}
}
\value{
A \code{shinyDesign} object with simulated count matrix stored in \code{simCounts}, and
updated spot metadata in \code{simcolData}.
}
\description{
This function generates simulated spatial gene expression counts by combining baseline domain-informative
expression patterns with perturbed spot locations. It allows scaling of sequencing depth and the
\code{sigma} parameter from the FG model, and can keep a proportion of genes undisturbed.
}
\examples{
\dontrun{
simulated_data <- simulation_Spatial(spadesign_obj, 
                                   seq_depth_factor = 1.5,
                                   SIGMA = 1.2,
                                   SEED = 123,
                                   prop = 0.7)
}
}
