% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_Spatial.r
\name{simulation_Spatial}
\alias{simulation_Spatial}
\title{Simulate data with modified sequencing depth and dispersion}
\usage{
simulation_Spatial(
  shinyDesign,
  selected_M_list = NULL,
  seq_depth_factor,
  SIGMA,
  SEED,
  prop
)
}
\arguments{
\item{shinyDesign}{A \code{shinyDesign} object}

\item{seq_depth_factor}{Numeric scaling factor for sequencing depth}

\item{SIGMA}{scaling factor for the \code{sigma} parameter from the FG model}

\item{SEED}{random seed for reproducibility}

\item{prop}{Proportion of genes to keep undisturbed}
}
\value{
A \code{shinyDesign} object with simulated count matrix
}
\description{
Simulate data with modified sequencing depth and dispersion
}
\examples{
\dontrun{
simulated_data <- simulation_Spatial(spadesign_obj, 
                                   seq_depth_factor = 1.5,
                                   SIGMA = 1.2,
                                   SEED = 123,
                                   prop = 0.7)
}
}
